feat: add logic for feed editing
