--- src/pages/post/[pid]/edit.tsx
@@ -1,13 +1,0 @@
-import NavLayout from "@/components/@common/NavLayout";
-
-import Head from "next/head";
-
-export default function PostEdit() {
-  return (
-    <NavLayout>
-      <Head>
-        <title>ArtBubble | Upload</title>
-      </Head>
-    </NavLayout>
-  );
-}
